add menus and seeders...

+++ routes/web.php
@@ -24,15 +24,15 @@ Route::get('/dashboard', function () {
     return view('dashboard');
 })->middleware(['auth', 'verified'])->name('dashboard');
 
-Route::get('/dashboard/admin', [TypeVehiculeController::class, 'index'])->name("admin");
-Route::get('/dashboard/admin/create', [TypeVehiculeController::class, 'create'])->name("admin.create");
-Route::post('/dashboard/admin/create', [TypeVehiculeController::class, 'store'])->name("admin.store");
+Route::get('/dashboard/admin', [TypeVehiculeController::class, 'index'])->middleware(['auth', 'verified'])->name("admin");
+Route::get('/dashboard/admin/create', [TypeVehiculeController::class, 'create'])->middleware(['auth', 'verified'])->name("admin.create");
+Route::post('/dashboard/admin/create', [TypeVehiculeController::class, 'store'])->middleware(['auth', 'verified'])->name("admin.store");
 
 //route pour l'édition
-Route::get('/dashboard/admin/{typeVehicule}', [TypeVehiculeController::class, 'edit'])->name("admin.edit");
-Route::put('/dashboard/admin/{typeVehicule}', [TypeVehiculeController::class, 'update'])->name("admin.update");
+Route::get('/dashboard/admin/{typeVehicule}', [TypeVehiculeController::class, 'edit'])->middleware(['auth', 'verified'])->name("admin.edit");
+Route::put('/dashboard/admin/{typeVehicule}', [TypeVehiculeController::class, 'update'])->middleware(['auth', 'verified'])->name("admin.update");
 //route pour la suppression
-Route::delete('/admin/{typeVehicule}', [TypeVehiculeController::class, 'delete'])->name("admin.delete");
+Route::delete('/admin/{typeVehicule}', [TypeVehiculeController::class, 'delete'])->middleware(['auth', 'verified']) ->name("admin.delete");
 Route::middleware('auth')->group(function () {
     Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
     Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
